dtdl in path, tests to update
