+++ .gitignore
@@ -11,3 +11,4 @@ obj
 /playwright-report/
 /playwright/.cache/
 *.zip
+.env
