+++ src/lib/server/utils/sessions.ts
@@ -9,7 +9,6 @@ export type Session = {
   search?: string
   highlightNodeId?: string
   expandedIds: string[]
-  dtdlModelId?: string
 }
 
 @singleton()
