+++ src/lib/server/views/components/mermaid.tsx
@@ -10,14 +10,14 @@ import { DtdlId, UUID } from '../../models/strings.js'
 import { getDisplayName, isInterface, isRelationship } from '../../utils/dtdl/extract.js'
 import { AccordionSection, Page } from '../common.js'
 
-const commonUpdateAttrs = {
+const commonUpdateAttrs = (dtdlModelId: UUID) => ({
   'hx-target': '#mermaid-output',
-  'hx-get': '/update-layout',
+  'hx-get': `update-layout`,
   'hx-swap': 'outerHTML  transition:true',
   'hx-include': '#sessionId, #search-panel',
   'hx-indicator': '#spinner',
   'hx-disabled-elt': 'select',
-}
+})
 
 function maybeNumberToAttr(value: number | undefined, defaultValue: number) {
   return `${value === undefined ? defaultValue : value}`
@@ -28,6 +28,7 @@ export default class MermaidTemplates {
   constructor() {}
 
   public MermaidRoot = ({
+    dtdlModelId,
     search,
     layout,
     sessionId,
@@ -35,6 +36,7 @@ export default class MermaidTemplates {
     svgWidth,
     svgHeight,
   }: {
+    dtdlModelId: UUID
     search?: string
     layout: Layout
     sessionId: UUID
@@ -46,6 +48,7 @@ export default class MermaidTemplates {
       <input id="sessionId" name="sessionId" type="hidden" value={escapeHtml(sessionId)} />
       <section id="toolbar">
         <this.searchPanel
+          dtdlModelId={dtdlModelId}
           layout={layout}
           search={search}
           diagramType={diagramType}
@@ -56,7 +59,7 @@ export default class MermaidTemplates {
       </section>
 
       <div id="mermaid-wrapper">
-        <this.mermaidTarget target="mermaid-output" />
+        <this.mermaidTarget target="mermaid-output" dtdlModelId={dtdlModelId} />
         <div id="spinner" />
       </div>
       <this.Legend showContent={false} />
@@ -88,15 +91,17 @@ export default class MermaidTemplates {
   public mermaidTarget = ({
     generatedOutput,
     target,
+    dtdlModelId,
   }: {
     generatedOutput?: JSX.Element
     target: string
+    dtdlModelId: UUID
   }): JSX.Element => {
     const attributes = generatedOutput
       ? { 'hx-on::after-settle': `globalThis.setMermaidListeners()`, 'pending-listeners': '' }
       : {
           'hx-trigger': 'load',
-          ...commonUpdateAttrs,
+          ...commonUpdateAttrs(dtdlModelId),
         }
     const output = generatedOutput ?? ''
     return (
@@ -210,6 +215,7 @@ export default class MermaidTemplates {
   }
 
   public searchPanel = ({
+    dtdlModelId,
     search,
     layout,
     swapOutOfBand,
@@ -221,6 +227,7 @@ export default class MermaidTemplates {
     currentPanY,
   }: {
     // inputs with current state
+    dtdlModelId: UUID
     search?: string
     layout: Layout
     diagramType: DiagramType
@@ -240,7 +247,7 @@ export default class MermaidTemplates {
         class="button-group"
         hx-swap-oob={swapOutOfBand ? 'true' : undefined}
         hx-sync="this:replace"
-        {...commonUpdateAttrs}
+        {...commonUpdateAttrs(dtdlModelId)}
       >
         <h2>UKDTC</h2>
         <input
@@ -250,7 +257,7 @@ export default class MermaidTemplates {
           value={escapeHtml(search || '')}
           placeholder="Search"
           hx-trigger="input changed delay:500ms, search"
-          {...commonUpdateAttrs}
+          {...commonUpdateAttrs(dtdlModelId)}
         />
 
         <input id="svgWidth" name="svgWidth" type="hidden" value={maybeNumberToAttr(svgWidth, 300)} />
@@ -260,7 +267,7 @@ export default class MermaidTemplates {
         <input id="currentPanY" name="currentPanY" type="hidden" value={maybeNumberToAttr(currentPanY, 0)} />
 
         <label for="diagramType">Diagram Type</label>
-        <select id="diagramType" name="diagramType" hx-trigger="input changed" {...commonUpdateAttrs}>
+        <select id="diagramType" name="diagramType" hx-trigger="input changed" {...commonUpdateAttrs(dtdlModelId)}>
           {diagramTypes.map((entry) => (
             <option value={entry} selected={entry === diagramType}>
               {escapeHtml(entry)}
@@ -273,7 +280,7 @@ export default class MermaidTemplates {
           name="layout"
           hx-trigger="input changed"
           disabled={diagramType === 'classDiagram'}
-          {...commonUpdateAttrs}
+          {...commonUpdateAttrs(dtdlModelId)}
         >
           {layoutEntries.map((entry) => (
             <option value={entry} selected={entry === layout}>
